recoded to use shapes_postgis
modifided routes/area_query_service to use shapes_postgis, and the
tempseg.tdetector table, which is a view that grabs both vds and wim detectors
in an area without regard to the time of use.  Also modified so that the current year in the query params is patched into the detector result, before passing on to couchdb getter.

modified package.json to add requirement to pull in shapes_postgis

committing in order to push over to anne for testing.