starting reversion to geoquery lib
getting all the detectors in an area and then pumping them all to a
couchdb query is a slow way to go.  Using geoQuery, you only get
features that actually have data.  So it is either faster, or the same
speed, in theory.

The psql query is probably slower

and I have to post sort the features to group by identical features

Anyway, this is the first commit on this trail.  Refactored the query
service, splitting out the json or csv handler.  added logging via
winston, although that is untested. added winston to package.json